+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Update `PlatformJavaVersions` and `PlatformKotlinVersions`
 - Introduce a separated Sandbox for the Frontend part when running IDE in Split Mode
 - Rename `SandboxAware.sandboxContainerDirectory` to `SandboxAware.sandboxDirectory` to avoid confusion with `intellijPlatform.sandboxContainer`
+- Use custom task name as a suffix for dynamically created configuration and tasks instead of `UUID.randomUUID()`
 
 ### Fixed